aws_fargate/gateway/apps/fossa_worker/requirements.txt
ayeaye-fossa
